add another parity example with event listener attached to a child div
this commit attempts to achieve parity between the components created
with the standard JS API and with Stencil. Here I add a div element to
the Stencil component as a child to the <Host> element and the parent
to the <input> and <button> elements. I moved the event listener from
the <Host> to the <div>, bringing it inline with the other component

In this commit, the DOM looks something like this:

```html
<body>
  <!-- Standard JS API created component, is unchanged from orignal reproduction case -->
  <blur-test>
    <div id="fake-host"> <!-- host event listener attached here -->
      <input id="input" placeholder="focus here first">
      <button>then click here</button>
    <div>
  </blur-test>
  <!-- Stencil component -->
  <my-component class="hydrated"> <!-- host event listener was attached here -->
    <div> <!-- host event listener is now attached here -->
      <input id="input" placeholder="focus here first">
      <button>then click here</button>
    <div>
  </my-component>
</body>
```