move event listener to custom element
the reproduction in this repository contains two custom elements for
the purpose compare a Stencil compiled component with one created via
standards JS APIs

comparing the two in FireFox dev tools, I noticed that the event
listener for the component created using JS APIs (not Stencil) had its
event listener attached to the div with the id 'fake-host', and not the
component itself (because <Host/> sets the event listener on the host
element itself). however, the Stencil version of this component has the
host blur event listener attached to the component itself. this commit
seeks to create parity between the two for evaluation purposes.

prior to the commit, the DOM looked something like:
```html
<body>
  <blur-test> <!-- standard API created component -->
    <div id="fake-host"> <!-- host blur event was attached here -->
      <input id="input" placeholder="focus here first">
      <button>then click here</button>
    </div>
  </blur-test>
  <!-- The Stencil component, 'my-component' -->
  <my-component class="hydrated"> <!-- host blur event was attached here -->
    <!-- Note, there's no parent div element here like in the example -->
    <input id="input" placeholder="focus here first">
    <button>then click here</button>
  </my-component>
</body>
```

after these modifications:
```html
<body>
  <!-- standard api created component -->
  <blur-test> <!-- host blur event is now attached here -->
    <div id="fake-host">
      <input id="input" placeholder="focus here first">
      <button>then click here</button>
    </div>
  </blur-test>
  <!-- The Stencil component, 'my-component' -->
  <my-component class="hydrated"> <!-- host blur event was attached here -->
    <!-- Note, there's no parent div element here like in the example -->
    <input id="input" placeholder="focus here first">
    <button>then click here</button>
  </my-component>
</body>
```